fix(security): use long arithmetic for skip+pageSize to resolve CodeQL integer-overflow finding
`(int) Math.min(skip + pageSize, total)` triggered CodeQL
"User-controlled data in arithmetic expression" because both `skip` and
`pageSize` are derived from request params. Cast to long before addition
matching the identical pattern already used at line 166 for the
predicates endpoint.

+++ backend/src/main/java/de/dlr/shepard/v2/semantic/resources/VocabularyBrowseRest.java
@@ -227,7 +227,7 @@ public class VocabularyBrowseRest {
     List<Vocabulary> used = vocabularyDAO.findVocabulariesUsedByEntity(entityAppId, scope);
     long total = used.size();
     int skip = (int) Math.min((long) page * pageSize, total);
-    List<Vocabulary> slice = used.subList(skip, (int) Math.min(skip + pageSize, total));
+    List<Vocabulary> slice = used.subList(skip, (int) Math.min((long) skip + pageSize, total));
     List<VocabularyIO> out = slice.stream().map(VocabularyIO::from).toList();
     return Response.ok(new PagedResponseIO<>(out, total, page, pageSize))
         .header("X-Total-Count", total)
